\name{baygel-package}
\alias{baygel-package}
\alias{baygel}
\docType{package}
\title{
  Bayesian Gaussian Graphical Model Estimation Library
}
\description{
 baygel provides estimation tools for Gaussian Graphical Models, with low to moderate dimensions, using 4 different Bayesian estimation methods. The estimators include, a Bayesian glasso; an adaptive Bayesian glasso; a ridge-type penalized estimator based on a Gaussian prior and finally an adaptive ridge-type penalized estimator based on an inverse-gamma generalisation of the Gaussian prior.

}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Your Name, email optional.

Maintainer: Your Name <your@email.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.
  }
}
