% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{blockBAGENII}
\alias{blockBAGENII}
\title{Type II naïve Bayesian adaptive graphical elastic net block Gibbs sampler for Gaussian graphical models.}
\usage{
blockBAGENII(X, burnin, iterations, verbose = TRUE, s = 1e-06, b = 0.5)
}
\arguments{
\item{X}{Numeric data matrix.}

\item{burnin}{An integer specifying the number of burn-in iterations.}

\item{iterations}{An integer specifying the length of the Markov chain after the burn-in iterations.}

\item{verbose}{A logical determining whether the progress of the MCMC sampler should be displayed.}

\item{s}{A double specifying the value of the rate parameter for the exponential prior associated with the Bayesian graphical lasso penalty term.}

\item{b}{A double specifying the value of the rate parameter for the exponential prior associated with the Bayesian graphical ridge penalty term.}
}
\value{
A list containing precision `Omega` and covariance `Sigma` matrices
from the Markov chains.
}
\description{
Implements the Type II naïve Bayesian adaptive graphical elastic net block Gibbs sampler to simulate the
posterior distribution of the precision matrix for Gaussian graphical models.
}
\examples{
# Generate true covariance matrix:
p             <- 10
n             <- 50
SigTrue       <- pracma::Toeplitz(c(0.7^rep(1:p-1)))
OmegaTrue     <- pracma::inv(SigTrue)
# Generate expected value vector:
mu            <- rep(0,p)
# Generate multivariate normal distribution:
set.seed(123)
X             <- MASS::mvrnorm(n, mu = mu, Sigma = SigTrue)
posterior     <- blockBAGENII(X, iterations = 1000, burnin = 500, b = 1e-3, s = 1e-1)
}
