% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{blockBAGR}
\alias{blockBAGR}
\title{Bayesian adaptive graphical ridge block Gibbs sampler for Gaussian graphical models.}
\usage{
blockBAGR(X, burnin, iterations, verbose = TRUE, a = 0.5, b = 0.5)
}
\arguments{
\item{X}{Numeric data matrix.}

\item{burnin}{An integer specifying the number of burn-in iterations.}

\item{iterations}{An integer specifying the length of the Markov chain after the burn-in iterations.}

\item{verbose}{A logical determining whether the progress of the MCMC sampler should be displayed.}

\item{a}{A double specifying the value of the shape parameter for the inverse gamma prior.}

\item{b}{A double specifying the value of the scale parameter for the inverse gamma prior.}
}
\value{
A list containing precision `Omega` and covariance `Sigma` matrices
from the Markov chains.
}
\description{
Implements a Bayesian adaptive graphical ridge block Gibbs sampler to simulate the
posterior distribution of the precision matrix for Gaussian graphical models.
}
\examples{
# Generate true covariance matrix:
p             <- 10
n             <- 50
SigTrue       <- pracma::Toeplitz(c(0.7^rep(1:p-1)))
OmegaTrue     <- pracma::inv(SigTrue)
# Generate expected value vector:
mu            <- rep(0,p)
# Generate multivariate normal distribution:
set.seed(123)
X             <- MASS::mvrnorm(n, mu = mu, Sigma = SigTrue)
posterior     <- blockBAGR(X, iterations = 1000, burnin = 500, a = 1, b = 1e-2)
}
