% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{blockBGEN}
\alias{blockBGEN}
\title{Block Gibbs sampler function.}
\usage{
blockBGEN(X, burnin, iterations, lambda = 1, sig = 1, verbose = TRUE)
}
\arguments{
\item{X}{Numeric data matrix, data is assumed to be Gaussian distributed.}

\item{burnin}{An integer specifying the number of burn-in iterations.}

\item{iterations}{An integer specifying the length of the Markov chain after the burn-in iterations.}

\item{lambda}{A numeric value representing the scale (rate) parameter for the double
exponential and exponential prior associated with the Bayesian graphical lasso penalty term.}

\item{sig}{A numeric value representing the standard deviation parameter for the double
Gaussian and truncated Gaussian prior associated with the Bayesian graphical ridge penalty term.}

\item{verbose}{A logical determining whether the progress of the MCMC sampler should be displayed.}
}
\value{
A list containing precision `Omega` and covariance `Sigma` matrices
from the Markov chains.
}
\description{
Implements the Bayesian graphical elastic net block Gibbs sampler to simulate the
posterior distribution of the concentration matrix in Gaussian graphical models.
}
\examples{
# Generate true covariance matrix:
p             <- 10
n             <- 50
SigTrue       <- pracma::Toeplitz(c(0.7^rep(1:p-1)))
OmegaTrue     <- pracma::inv(SigTrue)
# Generate expected value vector:
mu            <- rep(0,p)
# Generate multivariate normal distribution:
set.seed(123)
X             <- MASS::mvrnorm(n, mu = mu, Sigma = SigTrue)
posterior     <- blockBGEN(X, iterations = 1000, burnin = 500, lambda = 1, sig = 1)
}
