% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{blockBGL}
\alias{blockBGL}
\title{Bayesian graphical lasso block Gibbs sampler for Gaussian graphical models.}
\usage{
blockBGL(X, burnin, iterations, lambda = 1, verbose = TRUE)
}
\arguments{
\item{X}{A numeric matrix.}

\item{burnin}{An integer representing the number of burn-in iterations.}

\item{iterations}{An integer representing the length of the Markov chain post burn-in.}

\item{lambda}{A numeric value representing the scale (rate) parameter for the double
exponential and exponential prior.}

\item{verbose}{A logical indicating if the MCMC sampler progress should be printed.}
}
\value{
A list containing precision `Omega` and covariance `Sigma` matrices
from the Markov chains.
}
\description{
Implements a Bayesian graphical lasso block Gibbs sampler to simulate the
posterior distribution of the precision matrix for Gaussian graphical models.
}
\examples{
# Generate true covariance matrix:
p             <- 10
n             <- 50
SigTrue       <- pracma::Toeplitz(c(0.7^rep(1:p-1)))
OmegaTrue     <- pracma::inv(SigTrue)
# Generate expected value vector:
mu            <- rep(0,p)
# Generate multivariate normal distribution:
set.seed(123)
X             <- MASS::mvrnorm(n, mu = mu, Sigma = SigTrue)
posterior     <- blockBGL(X, iterations = 1000, burnin = 500, lambda = 0.5)
}
